backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
groq==0.11.0
tavily-python==0.5.0
supabase==2.9.1
apscheduler==3.10.4
python-dotenv==1.0.1
pydantic==2.9.0
sse-starlette==2.1.0
